More whiteness and some ideas
Some tweaks to the whiteness testing. It's not looking great.

Here's an idea for a metric: sheer number of white pixels in our canny
edge detected image. It should be a decent measure of sheer activity in
an image.

Also, metric comparisons. Do in parallel a whole bunch of different
comparisons varying both the metric and the parameters used and see
which performs best.